refactor(BcaDateTextField): move component to its own directory
