+++ src/pages/reportes/historic/index.tsx
@@ -12,7 +12,7 @@ import {
     useGetAllLevelsQuery,
 } from '~redux/api/bca-backend/reports/commonSlice'
 import EditToolbar from '~components/table/headers/toolbar'
-import BcaDateTextField from '~components/input/BcaDateTextField'
+import BcaDateTextField from '~/components/input/BcaDateTextField/BcaDateTextField'
 import ActualTable from '~components/reports/ActualTable'
 import { useAppSelector } from '~redux/hooks'
 import { downloadExcelFile } from '~/utils/download'
