+++ src/pages/reportes/spent/index.tsx
@@ -11,7 +11,7 @@ import {
     useGetSpentQuery,
 } from '~redux/api/bca-backend/reports/commonSlice'
 import BcaSelect from '~components/input/BcaSelect'
-import BcaDateTextField from '~components/input/BcaDateTextField'
+import BcaDateTextField from '~/components/input/BcaDateTextField/BcaDateTextField'
 import EditToolbar from '~components/table/headers/toolbar'
 import SpentTable from '~components/reports/SpentTable'
 import SpentDetailsDrawer from '~components/drawers/Reports/Spent/SpentDetailsDrawer'
